fix: add blocking deployments and jobs
This forces the deployments and jobs to block until the deployments is ready and
the jobs are complete. If you want to apply a configuration in the background it
can be put into a go routine. The main motivation for this is to fix debinding.
The secret that was needed in the the debinding jobs was getting deleted before
the debinding job container was booted